Integerリソースの扱いが一部間違っていたので修正

+++ type.h
@@ -1,12 +1,12 @@
 #ifndef _TYPE_H
 #define _TYPE_H
 
-typedef char int8;
-typedef int int16;
-typedef long int32;
-typedef long long int64;
+typedef signed char int8;
+typedef signed short int16;
+typedef signed long int32;
+typedef signed long long int64;
 typedef unsigned char uint8;
-typedef unsigned int uint16;
+typedef unsigned short uint16;
 typedef unsigned long uint32;
 typedef unsigned long long uint64;
 typedef float float32;
